feat(backend): contact controller, interface, service
The controller for contacts was created, as well as the interface and the service for obtaining the information, as well as their respective functions within them. The routing for the API was also created. In addition, a detail was fixed with the Address model, in which some data for the full_address attribute was incorrect